pyproject.toml
[project]
name = "planify-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyTelegramBotAPI>=4.15.0",
    "pymongo>=4.6.1",
    "google-genai>=0.3.0",
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
